refactor: move AgentHistoryManager
