+++ src/historyView/managers/index.ts
@@ -1,0 +1,1 @@
+export { AgentHistoryManager } from './AgentHistoryManager';
